refactor: app 패키지 바깥에 security 패키지를 선언
