+++ src/main/java/nextstep/app/ui/MemberController.java
@@ -2,7 +2,8 @@ package nextstep.app.ui;
 
 import nextstep.app.domain.Member;
 import nextstep.app.domain.MemberRepository;
-import nextstep.app.util.Base64Convertor;
+import nextstep.security.exception.AuthenticationException;
+import nextstep.security.util.Base64Convertor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestHeader;
