+++ src/main/java/nextstep/security/exception/AuthenticationException.java
@@ -1,8 +1,7 @@
-package nextstep.app.ui;
+package nextstep.security.exception;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
 
 @ResponseStatus(code = HttpStatus.UNAUTHORIZED)
-public class AuthenticationException extends RuntimeException {
-}
+public class AuthenticationException extends RuntimeException {}
