+++ src/test/java/nextstep/app/BasicAuthTest.java
@@ -2,7 +2,7 @@ package nextstep.app;
 
 import nextstep.app.domain.Member;
 import nextstep.app.domain.MemberRepository;
-import nextstep.app.util.Base64Convertor;
+import nextstep.security.util.Base64Convertor;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -42,8 +42,8 @@ class BasicAuthTest {
         String encoded = Base64Convertor.encode(token);
 
         ResultActions loginResponse = mockMvc.perform(get("/members")
-                .header("Authorization", "Basic " + encoded)
-                .contentType(MediaType.APPLICATION_FORM_URLENCODED_VALUE)
+                                                              .header("Authorization", "Basic " + encoded)
+                                                              .contentType(MediaType.APPLICATION_FORM_URLENCODED_VALUE)
         );
 
         loginResponse.andDo(print());
@@ -56,8 +56,8 @@ class BasicAuthTest {
     void members_fail() throws Exception {
         ResultActions loginResponse = mockMvc.
                 perform(get("/members")
-                        .header("Authorization", "Basic " + "invalid")
-                        .contentType(MediaType.APPLICATION_FORM_URLENCODED_VALUE)
+                                .header("Authorization", "Basic " + "invalid")
+                                .contentType(MediaType.APPLICATION_FORM_URLENCODED_VALUE)
                 );
 
         loginResponse.andDo(print());
